Quick Union generator (changed base and utils of QuickFind to be reused).
