+++ settings/questionsSelection.ts
@@ -23,7 +23,8 @@ import { RecursionFormula } from "@shared/question-generators/recursion/formula"
 import { RecurrenceMaster } from "@shared/question-generators/recursion/recurrenceMaster"
 import { stackQuestion } from "@shared/question-generators/Stack/StackGenerator.ts"
 import { Loops } from "@shared/question-generators/time/loops"
-import { QuickFindGenerator } from "@shared/question-generators/unionFind/quickFind/generatorQF"
+import { QuickFindGenerator } from "@shared/question-generators/unionFind/quickFind/generator.ts"
+import { QuickUnionGenerator } from "@shared/question-generators/unionFind/quickUnion/generator.ts"
 
 export const DEFAULT_IMAGE = new URL("../../assets/images/skill-default.jpg", import.meta.url)
 
@@ -84,7 +85,7 @@ export const collection: QuestionCollection = [
   {
     slug: "union-find",
     name: { de: "Union-Find", en: "Union-Find" },
-    contents: [QuickFindGenerator],
+    contents: [QuickFindGenerator, QuickUnionGenerator],
   },
 ]
