+++ shared/src/question-generators/unionFind/quickFind/algorithm.ts
