+++ shared/src/question-generators/unionFind/quickUnion/generator.ts
@@ -1,0 +1,53 @@
+import { QuestionGenerator } from "@shared/api/QuestionGenerator.ts"
+import { serializeGeneratorCall } from "@shared/api/QuestionRouter.ts"
+import { QuickUnion } from "@shared/question-generators/unionFind/quickUnion/algorithm.ts"
+import { unionFindStartQuestion } from "@shared/question-generators/unionFind/utilsStart.ts"
+import Random from "@shared/utils/random.ts"
+import { tFunctional, Translations } from "@shared/utils/translations.ts"
+
+const translations: Translations = {
+  en: {
+    name: "Quick-Union",
+    description: "Determine the Quick-Union state after Union operation",
+    task: `A state of the Quick-Union data structure is given as the following array: \n{{0}}\n We call \`Union({{1}}, {{2}})\`. Provide the resulting state. {{3}}
+      We assume that the operation **Union(**$i$**,** $j$**)** always sets the value specified by **Find(**$i$**)** to the value specified by **Find(**$j$**)**.`,
+  },
+  de: {
+    name: "Quick-Union",
+    description: "Bestimme den Quick-Union-Zustand nach Union-Operation",
+    task: `Ein Zustand der Quick-Union Datenstruktur ist als folgendes Array gegeben: \n{{0}}\n Wir rufen \`Union({{1}}, {{2}})\` auf. Gib den Zustand an, der dadurch entsteht. {{3}}
+      Wir nehmen an, dass die Operation **Union(**$i$**,** $j$**)** immer den durch **Find(**$i$**)** spezifizierten Wert auf den von **Find(**$j$**)** spezifierten Wert setzt.`,
+  },
+}
+
+export const QuickUnionGenerator: QuestionGenerator = {
+  id: "ufqu",
+  name: tFunctional(translations, "name"),
+  description: tFunctional(translations, "description"),
+  tags: ["union-find", "quick-find"],
+  languages: ["en", "de"],
+  expectedParameters: [
+    {
+      type: "string",
+      name: "variant",
+      allowedValues: ["start"],
+    },
+  ],
+
+  generate(lang, parameters, seed) {
+    const permalink = serializeGeneratorCall({
+      generator: QuickUnionGenerator,
+      lang,
+      parameters,
+      seed,
+    })
+    const random = new Random(seed)
+
+    const unionSize = random.int(7, 8)
+    const union = new QuickUnion(unionSize)
+
+    return {
+      question: unionFindStartQuestion({ random, union, unionSize, lang, permalink, translations }),
+    }
+  },
+}
